chatbot/pyproject.toml
[project]
name = "text2cypher-react-agent-example"
version = "0.1.0"
description = "A Text2Cypher ReAct agent built with LangGraph and the Neo4j Cypher MCP server."
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "geopy>=2.4.1",
    "fastapi>=0.115.0",
    "langchain>=0.3.27",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.29",
    "langgraph>=0.6.4",
    "neo4j>=5.28.2",
    "openai>=1.99.5",
    "pydantic>=2.11.7",
    "uvicorn>=0.31.0",
    "python-dotenv>=1.1.1",
     # Data manipulation and analysis
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.8",
    "pytest>=8.3.4",
]

[tool.ruff]
line-length = 100
